requirements-prod.txt
# Configuration + Utilities
hydra-core==1.3.2
omegaconf==2.3.0
python-json-logger==4.0.0
rich==14.2.0

# Machine Learning
opencv-python-headless==4.12.0.88
sentence-transformers==5.1.2
soundfile==0.13.1
transformers==4.57.1
ultralytics==8.3.230

# Database
pysqlite3==0.5.4

# Backend
fastapi==0.122.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
python-multipart==0.0.9
